+++ src/_messaging/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['_attachments', '_listening', '_reactions', '_send', '_unsend', '_message_requests', '_createNotes']
+__all__ = ['_attachments', '_changeTheme', '_createNotes', '_editMessage', '_listening', '_listening_e2ee', '_reactions', '_send', '_send_e2ee', '_unsend', '_message_requests']